requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.7
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.6
django-ninja==1.1.0
filelock==3.14.0
gunicorn==22.0.0
identify==2.5.36
idna==3.7
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.1
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.26.1
whitenoise==6.6.0
